requirements.txt
autopy
pillow
